tests/benchmarks/thread_metric: Don't select IRQ_OFFLOAD_NESTED
This is a mistake.  IRQ_OFFLOAD_NESTED is not intended to be an app
tunable. It reflects whether or not platform layer irq_offload()
implementation supports its use from within an ISR (i.e. to invoke a
nested interrupt for testing purposes).  Not all architectures can do
that (and I know of at least one Xtensa soc that didn't give a
software interrupt the right priority and can't either).  Basically
this should have been a "ARCH_HAS_" kind of kconfig, it's just named
funny.

And in any case the thread_metric tests don't do nested interrupts,
which as I understand it aren't even possible in the FreeRTOS
environment it was written on.

Signed-off-by: Andy Ross <andyross@google.com>